backend/requirements.txt
flask
flask-cors
fastapi
pydantic
tensorflow-cpu
numpy
scipy
transformers
torch
sentence-transformers
pydub
librosa
supabase
firebase-admin
python-dotenv